requirements.txt
psycopg2-binary==2.9.7
ibm_db==3.1.4
pandas==2.0.3
sqlalchemy==2.0.20
tabulate==0.9.0
colorama==0.4.6
pyyaml==6.0.1